+++ project/plugins.sbt
@@ -3,3 +3,4 @@ addSbtPlugin("ch.epfl.scala"  % "sbt-scalafix"   % "0.14.3")
 addSbtPlugin("com.github.sbt" % "sbt-dynver"     % "5.1.0")
 addSbtPlugin("com.github.sbt" % "sbt-ci-release" % "1.11.0")
 addSbtPlugin("org.typelevel"  % "sbt-tpolecat"   % "0.5.2")
+addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.19.0")
