--- workflows4s-web-ui/.scalafix.conf
@@ -1,22 +1,0 @@
-rules = [
-  DisableSyntax
-]
-
-OrganizeImports {
-  removeUnused = false
-}
-
-DisableSyntax {
-  noVars = true
-  noThrows = true
-  noNulls = true
-  noReturns = true
-  noWhileLoops = true
-  noAsInstanceOf = false
-  noIsInstanceOf = false
-  noXml = true
-  noDefaultArgs = true
-  noFinalVal = true
-  noFinalize = true
-  noValPatterns = true
-}
