--- workflows4s-web-ui/.scalafmt.conf
@@ -1,6 +1,0 @@
-version = 3.8.5
-runner.dialect = scala3
-style = defaultWithAlign
-maxColumn = 80
-rewrite.rules = [RedundantBraces,RedundantParens,PreferCurlyFors]
-danglingParentheses.preset = true
