--- workflows4s-web-ui/flake.nix
@@ -1,51 +1,0 @@
-{
-  description = "indigo-dev";
-
-  inputs = {
-    nixpkgs.url = github:nixos/nixpkgs/nixpkgs-unstable;
-    flake-utils.url = github:numtide/flake-utils;
-  };
-
-  outputs = { self, nixpkgs, flake-utils, ... }:
-    flake-utils.lib.eachDefaultSystem (system:
-      let
-        pkgs = import nixpkgs {
-          inherit system;
-          overlays = [
-            (f: p: {
-              mill = p.mill.override { jre = p.jdk17_headless; };
-              sbt = p.sbt.override { jre = p.jdk17_headless; };
-            })
-          ];
-        };
-        jdk = pkgs.jdk17_headless;
-
-        jvmInputs = with pkgs; [
-          jdk
-          coursier
-          mill
-          sbt
-        ];
-        jvmHook = ''
-          JAVA_HOME="${jdk}"
-        '';
-
-        jsInputs = with pkgs; [
-          nodejs
-          yarn
-          nodePackages_latest.http-server
-        ];
-        jsHook = ''
-          yarn install
-        '';
-      in
-      {
-        devShells.default = pkgs.mkShell {
-          name = "tyrian-dev-shell";
-          buildInputs = jvmInputs ++ jsInputs;
-          shellHook = jvmHook + jsHook;
-        };
-      }
-    );
-
-}
