--- workflows4s-web-ui/project/plugins.sbt
@@ -1,4 +1,0 @@
-addSbtPlugin("org.scala-js"        % "sbt-scalajs"  % "1.19.0")
-addSbtPlugin("org.typelevel"       % "sbt-tpolecat" % "0.5.2")
-addSbtPlugin("ch.epfl.scala"       % "sbt-scalafix" % "0.14.0")
-addSbtPlugin("com.github.reibitto" % "sbt-welcome"  % "0.5.0")
