--- workflows4s-web-ui/src/test/scala/myorg/workflows4s-web-uiTests.scala
@@ -1,9 +1,0 @@
-package myorg
-
-class workflows4s-web-uiTests extends munit.FunSuite {
-
-  test("dummy test") {
-    assert(1 == 1)
-  }
-
-}
